split phone and verto into audio/main sip-bridge and verto-bridge
